requirements.txt
beautifulsoup4==4.6.0
Flask==1.0.2
gunicorn==19.9.0
line-bot-sdk==1.8.0

